+++ frontend/src/app/features/main-page/components/description-section/description-section.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, OnInit, ChangeDetectionStrategy } from '@angular/core';
+import { Component, Input, ChangeDetectionStrategy, EventEmitter, Output } from '@angular/core';
 import { MatTooltip } from '@angular/material/tooltip';
 
 import { Game } from './../../../../interfaces/game.interface';
@@ -13,19 +13,22 @@ import { GameCategory } from '../../../../interfaces/game-category.interface';
     styleUrls: ['./description-section.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class DescriptionSectionComponent implements OnInit {
+export class DescriptionSectionComponent {
     @Input()
-    public description: DescriptionResponseBody;
+    public set descriptionData(description: DescriptionResponseBody) {
+        this.description = description;
+        this.prepareDescriptionToCopy();
+    };
 
+    @Output()
+    public reloadDescription: EventEmitter<void> = new EventEmitter<void>();
+
+    public description: DescriptionResponseBody;
     public descriptionToCopy: string;
     public sectionSeparator: string = '_______________________________';
     public tooltipText: string = 'Copied!';
     public headingText: string = 'Description';
     public shouldShowEdition: boolean;
-
-    public ngOnInit(): void {
-        this.prepareDescriptionToCopy();
-    }
 
     public handleEditionMode(): void {
         this.shouldShowEdition = !this.shouldShowEdition;
@@ -41,6 +44,10 @@ export class DescriptionSectionComponent implements OnInit {
         }, 2000);
     }
 
+    public deleteGame(): void {
+        this.reloadDescription.emit();
+    }
+
     private prepareDescriptionToCopy(): void {
         let descriptionToCopy: string = '';
 
@@ -48,13 +55,17 @@ export class DescriptionSectionComponent implements OnInit {
         descriptionToCopy += this.getCategoriesPartOfDescription(this.description?.categories);
 
         // Separator 1
-        descriptionToCopy += this.getSeparatorPartOfDescription('Completed', this.description?.completedGames?.length);
+        if (this.description.completedGames?.length) {
+            descriptionToCopy += this.getSeparatorPartOfDescription('Completed', this.description.completedGames?.length);
+        }
 
         // Completed games
         descriptionToCopy += this.getGamePartOfDescription(this.description?.completedGames);
 
         // Separator 2
-        descriptionToCopy += this.getSeparatorPartOfDescription('New achievements', this.description?.gamesWithNewAchievements?.length);
+        if (this.description.gamesWithNewAchievements?.length) {
+            descriptionToCopy += this.getSeparatorPartOfDescription('New achievements', this.description.gamesWithNewAchievements?.length);
+        }
 
         // Games with new achievements
         descriptionToCopy += this.getGamePartOfDescription(this.description?.gamesWithNewAchievements);