+++ frontend/src/app/features/main-page/main-page.component.ts
@@ -7,7 +7,7 @@ import { EMPTY, Observable, of, zip } from 'rxjs';
 import { DescriptionResponseBody } from './../../interfaces/description-response-body.interface';
 import { ViewState } from '../../enums/view-state.enum';
 import { Destroyable } from '../../utils/destroyable.util';
-import { CategoriesResponseBody } from './../../interfaces/categories-response-body.interface';
+import { Category } from './../../interfaces/category.interface';
 
 @Component({
     selector: 'mados-main-page',
@@ -15,10 +15,11 @@ import { CategoriesResponseBody } from './../../interfaces/categories-response-b
     styleUrls: ['./main-page.component.scss'],
 })
 export class MainPageComponent extends Destroyable implements OnInit {
-    public viewState: ViewState;
+    public viewState: ViewState = ViewState.LOADING;
     public ViewState: typeof ViewState = ViewState;
-    public categories$: Observable<CategoriesResponseBody>;
+    public categories$: Observable<Category[]>;
     public description$: Observable<DescriptionResponseBody>;
+    public infobarMessage: string = 'To add a game and see the description, please fulfill all categories';
 
     constructor(private httpClient: HttpClient) {
         super();
@@ -28,40 +29,33 @@ export class MainPageComponent extends Destroyable implements OnInit {
         this.initCategoriesAndDescriptionData();
     }
 
-    public reloadDescription(): void {
+    public getDescriptionData(): void {
         this.httpClient.get<DescriptionResponseBody>('description')
             .pipe(
-                catchError(() => {
-                    // TODO: Add snackbar
-
-                    return EMPTY;
-                }),
                 map((description: DescriptionResponseBody) => {
                     this.description$ = of(description);
                 }),
                 takeUntil(this.destroyed$)
             )
-            .subscribe(() => {
-                // TODO Add snackbar
-            });
+            .subscribe();
     }
 
     private initCategoriesAndDescriptionData(): void {
-        this.viewState = ViewState.LOADING;
-
-        zip(this.httpClient.get<CategoriesResponseBody>('categories'), this.httpClient.get<DescriptionResponseBody>('description'))
+        zip(this.httpClient.get<Category[]>('categories'), this.httpClient.get<DescriptionResponseBody>('description'))
             .pipe(
                 catchError(() => {
                     this.viewState = ViewState.ERROR;
 
                     return EMPTY;
                 }),
-                map(([categories, description]: [CategoriesResponseBody, DescriptionResponseBody]) => {
+                map(([categories, description]: [Category[], DescriptionResponseBody]) => {
                     this.categories$ = of(categories);
                     this.description$ = of(description);
                 }),
                 takeUntil(this.destroyed$)
             )
-            .subscribe(() => (this.viewState = ViewState.SUCCESS));
+            .subscribe(() => {
+                this.viewState = ViewState.SUCCESS;
+            });
     }
 }