+++ frontend/src/app/interfaces/game.interface.ts
@@ -2,6 +2,7 @@ import { GameCategory } from './game-category.interface';
 
 export interface Game {
     id: number;
+    hasNewAchievements: boolean;
     oneHundredPercentIconName: string;
     name: string;
     categories?: GameCategory[];